+++ src/inspector/components/Inspector/Inspector.tsx
@@ -4,8 +4,9 @@ import { TabPanel } from "../Tabs/TabPanel";
 import "./../../../shared/ThemeVars.css";
 import "./Inspector.less";
 import { LeftColumnContainer } from "../LeftColumn/LeftColumnContainer";
-import { TActiveSection, TActiveInspectorTab } from "../../model/types";
+import { TActiveSection, TActiveInspectorTab, IDescriptor } from "../../model/types";
 import { FooterContainer } from "../Footer/FooterContainer";
+import { VisualDiffTab } from "../VisualDiff/VisualDiff";
 
 export interface IInspectorProps{
 	mainTab: TActiveSection
@@ -13,6 +14,7 @@ export interface IInspectorProps{
 	descriptorContent: string
 	calculatedReference: string
 	originalReference: string
+	selectedDescriptors: IDescriptor[]
 }
 
 export interface IInspectorDispatch {
@@ -45,14 +47,16 @@ export class Inspector extends React.Component<TInspector> {
 									</div>
 								</TabPanel>
 								<TabPanel id="difference" title="Difference" >
-									Difference
+									<div className="diff">
+										<VisualDiffTab left={this.props.selectedDescriptors[0]?.originalData} right={this.props.selectedDescriptors[1]?.originalData} />
+									</div>
 								</TabPanel>
 								<TabPanel id="reference" title="Info" >
 									<div className="info code">
-										Reference:
-										{this.props.calculatedReference}
-										
 										Filter:
+										{this.props.calculatedReference}
+										<br/>
+										Reference:
 										{this.props.originalReference}
 									</div>
 								</TabPanel>